+++ polytope_server/common/authentication/authentication.py
@@ -77,6 +77,7 @@ type_to_class_map = {
     "plain": "PlainAuthentication",
     "keycloak": "KeycloakAuthentication",
     "federation": "FederationAuthentication",
+    "jwt" : "JWTBearerAuthentication",
 }
